api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1

# Supabase v2 (required for storage + db)
supabase==2.4.5
postgrest==0.16.4
storage3==0.7.4

# OpenAI
openai==1.12.0

# HTTP
httpx==0.27.0

# File upload
python-multipart==0.0.9
